fix(count_min): use correct type size for seeds in size_bytes()

+++ src/frequency/count_min.rs
@@ -234,7 +234,7 @@ impl Sketch for CountMinSketch {
     fn size_bytes(&self) -> usize {
         core::mem::size_of::<Self>()
             + self.depth * self.width * core::mem::size_of::<u64>()
-            + self.seeds.len() * core::mem::size_of::<u32>()
+            + self.seeds.len() * core::mem::size_of::<u64>()
     }
 
     fn count(&self) -> u64 {
